+++ client/python/projectairsim/src/projectairsim/__init__.py
@@ -11,6 +11,7 @@ from .rover import Rover
 from .env_actor import EnvActor
 from .static_sensor_actor import StaticSensorActor
 from .unreal_vehicle import UnrealVehicle
+from .wheeled_vehicle import WheeledVehicle
 
 __version__ = "1.0.2"
 
@@ -22,4 +23,5 @@ __all__ = [
     "EnvActor",
     "StaticSensorActor",
     "UnrealVehicle",
+    "WheeledVehicle",
 ]